Separate functional and unit tests
